♻️ refactor : 공통 인스턴스 사용으로 변경
